Merge 'documents' with 'master'

+++ NSI.BLL/DocumentManipulation.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using NSI.BLL.Interfaces;
 using NSI.DC.DocumentRepository;
@@ -17,44 +19,68 @@ namespace NSI.BLL
         }
 
 
-        public List<DocumentDto> GetAllDocuments()
+        public List<DocumentDetails> GetAllDocuments()
         {
             return _documentRepository.GetAllDocuments();
         }
 
-        public void UploadFile(List<IFormFile> files, string filePath)
+        public async Task<string> UploadFileAsync(IFormFile file)
         {
-            var document = new DocumentDto()
+
+            if (file == null || file.Length == 0) return "File not selected";
+
+            var path = Path.Combine(
+                Directory.GetCurrentDirectory(), "Documents",
+                file.FileName);
+
+            using (var stream = new FileStream(path, FileMode.Create))
             {
-                DocumentPath = filePath
-            };
-            //TODO
+                await file.CopyToAsync(stream);
+            }
+            return path;
         }
 
         public bool SaveDocument(DocumentDto document)
         {
+            if (document == null) throw new Exception("Document is null");
+
             var result = _documentRepository.SaveDocument(document);
             return result != null;
         }
 
-        public PagingResultModel<DocumentDto> GetDocumentsByPage(DocumentsPagingQueryModel query)
+        public List<DocumentHistoryDto> GetDocumentHistoryByDocumentId(int id)
+        {
+            if (id == 0) throw new Exception("Id is not valid");
+            return _documentRepository.GetDocumentHistoryByDocumentId(id);
+        }
+
+        public List<DocumentDto> GetDocumentsByCase(int id)
+        {
+            return _documentRepository.GetDocumentsByCase(id);
+        }
+
+        public PagingResultModel<DocumentDetails> GetDocumentsByPage(DocumentsPagingQueryModel query)
         {
             if(query.PageNumber < 0) throw new Exception("Page number is not valid");
             return _documentRepository.GetAllDocumentsByPage(query);
         }
 
-        public DocumentDto GetDocumentById(int documentId)
+        public DocumentDetails GetDocumentById(int documentId)
         {
+            if (documentId == 0) throw new Exception("Id is not valid.");
             return _documentRepository.GetDocument(documentId);
         }
 
         public bool DeleteDocument(int id)
         {
+            if (id == 0) throw new Exception("Id is not valid");
             return _documentRepository.DeleteDocument(id);
         }
 
-        public bool EditDocument(DocumentDto documentDto)
+        public bool EditDocument(int id, DocumentDto documentDto)
         {
+            if (id == 0) throw new Exception("Id is not valid");
+
             var document = _documentRepository.GetDocument(documentDto.DocumentId);
             if (document == null) return false;
